added yellow schema

+++ week_4_analytics_engineering/web_to_gcs.py
@@ -48,20 +48,20 @@ table_schema_yellow = pa.schema(
         ('tpep_pickup_datetime', pa.timestamp('s')), 
         ('tpep_dropoff_datetime', pa.timestamp('s')), 
         ('passenger_count', pa.int64()), 
-        ('trip_distance', pa.int64()), 
+        ('trip_distance', pa.float64()), 
         ('RatecodeID', pa.string()), 
         ('store_and_fwd_flag', pa.string()), 
         ('PULocationID', pa.int64()), 
         ('DOLocationID', pa.int64()), 
         ('payment_type', pa.int64()), 
-        ('fare_amount',pa.int64()), 
-        ('extra',pa.int64()), 
-        ('mta_tax', pa.int64()), 
-        ('tip_amount', pa.int64()), 
-        ('tolls_amount', pa.int64()), 
-        ('improvement_surcharge', pa.int64()), 
-        ('total_amount', pa.int64()), 
-        ('congestion_surcharge', pa.int64())]
+        ('fare_amount',pa.float64()), 
+        ('extra',pa.float64()), 
+        ('mta_tax', pa.float64()), 
+        ('tip_amount', pa.float64()), 
+        ('tolls_amount', pa.float64()), 
+        ('improvement_surcharge', pa.float64()), 
+        ('total_amount', pa.float64()), 
+        ('congestion_surcharge', pa.float64())]
 
 )
